% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckeye_points.R
\docType{data}
\name{buckeye_points}
\alias{buckeye_points}
\title{Buckeye Formant Data}
\format{
A data frame with 225722 rows and 54 variables:
\describe{
\item{\code{age, city, ethnicity, first_name, last_name, location,
              state, year, years_of_schooling, style}}{
character These are all data columns that could
potentially be defined when calling FAVE-extract,
but in this case were not.
}
\item{\code{name}}{character Speaker name, in this case, the Buckeye corpus ID}
\item{\code{sex}}{character Speaker gender, as provided by
the Buckeye corpus handbook}
\item{\code{tiernum}}{double Speaker tier}
\item{\code{vowel}}{character CMUDICT vowel code}
\item{\code{stress}}{character vowel stress}
\item{\code{pre_word, word, fol_word}}{
character Orthographic transcriptions  of the preceding, current,
and following words
}
\item{\code{F1, F2, F2}}{double First, second and third formants in Hz.}
\item{\code{B1, B2, B3}}{double The bandwidths of F1, F2 and F3}
\item{\code{t}}{
double The measurement point for the vowel, in seconds, relative to
the entire recording.
}
\item{\code{beg}}{
double Vowel start time, in seconds, relative to the entire recording.
}
\item{\code{end}}{
double Vowel end time, in seconds, relative to the entire recording.
}
\item{\code{dur}}{
double Vowel duration, in seconds. Equal to \code{end - beg}
}
\item{\code{plt_vclass}}{character COLUMN_DESCRIPTION}
\item{\code{ipa_vclass}}{character COLUMN_DESCRIPTION}
\item{\code{plt_manner}}{character COLUMN_DESCRIPTION}
\item{\code{plt_place}}{character COLUMN_DESCRIPTION}
\item{\code{plt_voice}}{character COLUMN_DESCRIPTION}
\item{\code{plt_preseg}}{character COLUMN_DESCRIPTION}
\item{\code{plt_folseq}}{character COLUMN_DESCRIPTION}
\item{\code{glide}}{character COLUMN_DESCRIPTION}
\item{\code{pre_seg}}{character COLUMN_DESCRIPTION}
\item{\code{fol_seg}}{character COLUMN_DESCRIPTION}
\item{\code{context}}{character COLUMN_DESCRIPTION}
\item{\code{vowel_index}}{double COLUMN_DESCRIPTION}
\item{\code{pre_word_trans}}{character COLUMN_DESCRIPTION}
\item{\code{word_trans}}{character COLUMN_DESCRIPTION}
\item{\code{fol_word_trans}}{character COLUMN_DESCRIPTION}
\item{\code{F1@20\%}}{double COLUMN_DESCRIPTION}
\item{\code{F2@20\%}}{double COLUMN_DESCRIPTION}
\item{\code{F1@35\%}}{double COLUMN_DESCRIPTION}
\item{\code{F2@35\%}}{double COLUMN_DESCRIPTION}
\item{\code{F1@50\%}}{double COLUMN_DESCRIPTION}
\item{\code{F2@50\%}}{double COLUMN_DESCRIPTION}
\item{\code{F1@65\%}}{double COLUMN_DESCRIPTION}
\item{\code{F2@65\%}}{double COLUMN_DESCRIPTION}
\item{\code{F1@80\%}}{double COLUMN_DESCRIPTION}
\item{\code{F2@80%}}{double COLUMN_DESCRIPTION}
  \item{\code{nFormants}}{double COLUMN_DESCRIPTION}
}
}
\usage{
buckeye_points
}
\description{
All FAVE-extrated formant data from the Buckeye Corpus
}
\details{
DETAILS
}
\keyword{datasets}
