% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckeyefave.R
\docType{package}
\name{buckeyefave}
\alias{buckeyefave}
\title{\code{buckeyefave}: A data package for data derived from the Buckeye Corpus}
\description{
This is a data package containing the result of force-aligning the
\href{https://buckeyecorpus.osu.edu/}{Buckeye Corpus} (Pitt et al 2007) with
the \href{https://montreal-forced-aligner.readthedocs.io/en/latest/}{Montreal Forced Aligner}
and doing formant extraction with \href{https://github.com/joFrhwld/FAVE/}{FAVE}.
}
\references{
Pitt, M.A., Dilley, L., Johnson, K., Kiesling, S., Raymond, W., Hume, E.
and Fosler-Lussier, E. (2007) Buckeye Corpus of Conversational Speech
(2nd release) \link{www.buckeyecorpus.osu.edu} Columbus, OH: Department of
Psychology, Ohio State University (Distributor).
}
